apps/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile ../requirements/requirements.in
#
asgiref==3.7.2
    # via
    #   daphne
    #   django
attrs==23.1.0
    # via
    #   -r ../requirements/requirements.in
    #   automat
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
cffi==1.15.1
    # via cryptography
constantly==15.1.0
    # via twisted
contourpy==1.1.0
    # via matplotlib
contrib==0.3.0
    # via -r ../requirements/requirements.in
cryptography==41.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
cycler==0.11.0
    # via matplotlib
daphne==4.0.0
    # via -r ../requirements/requirements.in
django==4.2.3
    # via
    #   -r ../requirements/requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r ../requirements/requirements.in
fonttools==4.41.1
    # via matplotlib
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   hyperlink
    #   twisted
incremental==22.10.0
    # via twisted
jsonschema==4.18.4
    # via contrib
jsonschema-specifications==2023.7.1
    # via jsonschema
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.7.2
    # via contrib
numpy==1.25.1
    # via
    #   contourpy
    #   matplotlib
packaging==23.1
    # via matplotlib
pillow==10.0.0
    # via matplotlib
psycopg2==2.9.6
    # via -r ../requirements/requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyopenssl==23.2.0
    # via twisted
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   contrib
    #   matplotlib
pytz==2023.3
    # via djangorestframework
pyyaml==6.0.1
    # via contrib
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
service-identity==23.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   python-dateutil
sqlparse==0.4.4
    # via django
twisted[tls]==22.10.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.7.1
    # via twisted
zope-interface==6.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
